+++ contracts/Dependencies/ProbityMath.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 import "./SafeMath.sol";
 
 library ProbityMath {
-  using SafeMath for uint;
+  using SafeMath for uint256;
 
-  uint internal constant DECIMAL_PRECISION = 1e18;
+  uint256 internal constant DECIMAL_PRECISION = 1e18;
 }
