+++ contracts/Interfaces/IExchange.sol
@@ -3,8 +3,14 @@
 pragma solidity ^0.8.0;
 
 interface IExchange {
-  
   // --- Events ---
 
-  function executeOrder(address lender, address borrower, uint amount, uint rate) external;
+  // --- Functions ---
+
+  function executeOrder(
+    address lender,
+    address borrower,
+    uint256 amount,
+    uint256 rate
+  ) external;
 }
