+++ contracts/Interfaces/ITeller.sol
@@ -6,13 +6,22 @@ pragma solidity ^0.8.0;
  * @notice Manages debts for all vaults.
  */
 interface ITeller {
-
   // --- Events ---
 
-  event LoanCreated(address lender, address borrower, uint principal, uint rate, uint _now);
+  event LoanCreated(
+    address lender,
+    address borrower,
+    uint256 principal,
+    uint256 rate,
+    uint256 _now
+  );
 
   // --- Functions ---
 
-  function createLoan(address lender, address borrower, uint principal, uint rate) external;
-
+  function createLoan(
+    address lender,
+    address borrower,
+    uint256 principal,
+    uint256 rate
+  ) external;
 }
